requirements.txt
pandas
sqlalchemy
psycopg2-binary
streamlit
requests
nltk
transformers